+++ regimes/pt/migrations_test.go
@@ -54,7 +54,7 @@ func TestTaxZoneMigration(t *testing.T) {
 					Zone:    "20", //nolint:staticcheck
 				},
 			},
-			region: "PT-AC",
+			region: "PT",
 		},
 		{
 			name: "Madeira zone set",
@@ -62,25 +62,6 @@ func TestTaxZoneMigration(t *testing.T) {
 				TaxID: &tax.Identity{
 					Country: "PT",
 					Zone:    "30", //nolint:staticcheck
-				},
-			},
-			region: "PT-MA",
-		},
-		{
-			name: "Other zone set",
-			supplier: &org.Party{
-				TaxID: &tax.Identity{
-					Country: "PT",
-					Zone:    "40", //nolint:staticcheck
-				},
-			},
-			region: "PT",
-		},
-		{
-			name: "No zone set",
-			supplier: &org.Party{
-				TaxID: &tax.Identity{
-					Country: "PT",
 				},
 			},
 			region: "PT",